+++ amiga-mui/iconclass.c
@@ -252,7 +252,7 @@ STATIC ULONG Icon_Setup(struct IClass *cl, Object *obj, struct MUIP_Setup *msg)
 	data->ehnode.ehn_Class    = cl;
 	data->ehnode.ehn_Events   = IDCMP_MOUSEBUTTONS;
 
-	DoMethod(_win(obj), MUIM_Window_AddEventHandler, &data->ehnode);
+	DoMethod(_win(obj), MUIM_Window_AddEventHandler, (ULONG)&data->ehnode);
 
 	return 1;
 }
@@ -263,7 +263,7 @@ STATIC ULONG Icon_Setup(struct IClass *cl, Object *obj, struct MUIP_Setup *msg)
 STATIC ULONG Icon_Cleanup(struct IClass *cl, Object *obj, Msg msg)
 {
 	struct Icon_Data *data = (struct Icon_Data*)INST_DATA(cl,obj);
-	DoMethod(_win(obj), MUIM_Window_RemEventHandler, &data->ehnode);
+	DoMethod(_win(obj), MUIM_Window_RemEventHandler, (ULONG)&data->ehnode);
 	if (data->obj)
 	{
 		FreeDiskObject(data->obj);
@@ -370,9 +370,9 @@ STATIC ULONG Icon_HandleEvent(struct IClass *cl, Object *obj, struct MUIP_Handle
 
 			if (WorkbenchBase->lib_Version >= 45)
 			{
-				DoMethod(_win(obj), MUIM_Window_RemEventHandler, &data->ehnode);
+				DoMethod(_win(obj), MUIM_Window_RemEventHandler, (ULONG)&data->ehnode);
 				data->ehnode.ehn_Events |= IDCMP_MOUSEMOVE;
-				DoMethod(_win(obj), MUIM_Window_AddEventHandler, &data->ehnode);
+				DoMethod(_win(obj), MUIM_Window_AddEventHandler, (ULONG)&data->ehnode);
 			}
 
 			return MUI_EventHandlerRC_Eat;
@@ -382,9 +382,9 @@ STATIC ULONG Icon_HandleEvent(struct IClass *cl, Object *obj, struct MUIP_Handle
 		{
 			if (WorkbenchBase->lib_Version >= 45)
 			{
-				DoMethod(_win(obj), MUIM_Window_RemEventHandler, &data->ehnode);
+				DoMethod(_win(obj), MUIM_Window_RemEventHandler, (ULONG)&data->ehnode);
 				data->ehnode.ehn_Events &= ~IDCMP_MOUSEMOVE;
-				DoMethod(_win(obj), MUIM_Window_AddEventHandler, &data->ehnode);
+				DoMethod(_win(obj), MUIM_Window_AddEventHandler, (ULONG)&data->ehnode);
 			}
 			return MUI_EventHandlerRC_Eat;
 		}
@@ -569,16 +569,16 @@ STATIC ULONG Icon_DeleteDragImage(struct IClass *cl, Object *obj, Msg msg)
 
 				if (data->drop_path)
 				{
-					DoMethod(_app(obj),MUIM_Application_PushMethod, obj, 3, MUIM_Set, MUIA_Icon_DropPath, data->drop_path);
+					DoMethod(_app(obj),MUIM_Application_PushMethod, (ULONG)obj, 3, MUIM_Set, MUIA_Icon_DropPath, (ULONG)data->drop_path);
 				}
 			}
 		}
 	}
 #endif
 
-	DoMethod(_win(obj), MUIM_Window_RemEventHandler, &data->ehnode);
+	DoMethod(_win(obj), MUIM_Window_RemEventHandler, (ULONG)&data->ehnode);
 	data->ehnode.ehn_Events &= ~IDCMP_MOUSEMOVE;
-	DoMethod(_win(obj), MUIM_Window_AddEventHandler, &data->ehnode);
+	DoMethod(_win(obj), MUIM_Window_AddEventHandler, (ULONG)&data->ehnode);
 
 	return DoSuperMethodA(cl,obj,(Msg)msg);
 }
