+++ amiga-mui/muistuff.c
@@ -195,7 +195,7 @@ VOID DisposeAllChilds(Object *o)
 
   while ((child = (Object*)NextObject(&cstate)))
   {
-    DoMethod(o,OM_REMMEMBER,child);
+    DoMethod(o,OM_REMMEMBER, (ULONG)child);
     MUI_DisposeObject(child);
   }
 }
@@ -208,7 +208,7 @@ VOID DisposeAllFamilyChilds(Object *o)
 
   while ((child = (Object*)NextObject(&cstate)))
   {
-    DoMethod(o,OM_REMMEMBER,child);
+    DoMethod(o,OM_REMMEMBER, (ULONG)child);
     MUI_DisposeObject(child);
   }
 }
